Moved API views into separate blueprint.
